Remove `activate-user` deprecation warning
`nh` isn't updated for the modern config yet (well, how can it, it's not
released yet), and it's annoying to see the deprecation warning on every
rebuild. And the added features by `nh` are really cool and I don't
wanna stop using it (or go back to my old script setup).