treewide: specify full path to home directories in `/var/lib`
It seems macOS will resolve the home directory paths on system upgrades
(it happened on the 15.7 upgrade for me), which will break system
activation unless the user then manually changes the path in their
configuration or sets it back for each user. So set it as the default to
be compatible with updates.

In both cases manual intervention will be required sooner or later,
either with this commit to adjust the home directories of existing users
to their full paths (on systems that are not yet upgraded), or without
it, to change the configuration to reference the full paths manually
after upgrading.

So it's kind of a breaking change, in that activation will fail with an
error for existing users, but it will fail if they upgrade their system
anyway. The only case in which there will be no failure for existing
installations, is both if the system is upgraded, and only after the
upgrade this commit is applied. But since `/var` is a symlink to
`/private/var`, the data remains in exactly the same place, so no manual
data migration is required.